backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0
anthropic>=0.40.0
boto3>=1.35.0
redis>=5.0.0
celery>=5.4.0
pgvector>=0.3.0
python-dotenv>=1.0.0
